requirements.txt
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
beautifulsoup4==4.13.4
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.2
constantly==23.10.4
cryptography==44.0.3
daphne==4.1.2
defusedxml==0.7.1
Django==4.2.11
djangorestframework==3.16.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyOpenSSL==25.0.0
requests==2.32.3
service-identity==24.2.0
setuptools==80.4.0
soupsieve==2.7
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
urllib3==2.4.0
watchdog==6.0.0
whitenoise==6.9.0
zope.interface==7.2
